Wrapped RSVP.promise resolve in run.next loop
This is to enable the use of ember rsvp promises in ember-concurrency.  For some reason, Ember.rsvp promises do not yield at all as expected in ember concurrency and can only do so when promise's resolve method is called in the next run loop which is the update applied by this commit.